Move out the HPS specific code, mostly, to separate script
