+++ RunData/MyaData.py
@@ -3,6 +3,7 @@ import requests
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
 requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
 
+import pandas as pd
 from datetime import datetime,timedelta
 
 class MyaData:
